wsd: migrate Util::JsonToMap to JsonUtil::jsonToMap
