+++ kit/KitHelper.hpp
@@ -87,7 +87,7 @@ namespace LOKitHelper
                 ptrValue = loKitDocument->pClass->getPartInfo(loKitDocument, i);
                 const std::string partinfo(ptrValue);
                 std::free(ptrValue);
-                for (const auto& prop : Util::JsonToMap(partinfo))
+                for (const auto& prop : JsonUtil::jsonToMap(partinfo))
                 {
                     const std::string& name = prop.first;
                     if (name == "visible")
